Fixed Pages Nav
Fixed the issue with the sign in blocks giving problems. Damn programming languages love to use "end" as a reserved name and not tell you that. Format is now

<%=render 'shared/start'%>
<%=render 'shared/about'%>
<%=render 'shared/last'%>

Start checks for sign in. If true, returns profile name and link. If false, returns login/register link. Links to toggle the next block will be added in due course

About is meant to be swappable. Provides context dependent links, e.g. "subjects" or "about us" related.

Last checks for sign in. If true, returns a log out link. If false, returns nothing.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>