Fixed Submit Link
Use JS to find form name="myform" and submit that form. Since there'll be only one form on each page it works. A little hackish - ideally one would identify by the id of the form on that page… but that would mean parsing some variable somewhere. This should be easier.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>