Fixed Pacman
Turns out the problem was with the audio files.

Added the audio files, then fixed the path within pacman.js (line 1060) to link correctly. Didn't use the rails asset tag so modernizr can swap out the extension as appropriate depending on browser support.

Ran rake asset:clean to rebuild the JS because the previous version was persistent in the cache.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>