Whole bunch of format updates (See Description)
Let's see…

Trimmed utilities.css for useless classes

Linked the login template to the same partial so the formatting is consistent.

Devise shared links goes to hell as a consequence

Simpleforms goes to hell because it tries to act smart and fails

Added the school autocomplete to the staff sign up form

Moved the reveal style out of the 404 handler and into the reveal template

Abstracted application and reveal layouts for consistency and ease of management.

Moved the bootstrap flash helper and GCF into div.container so it obeys responsiveness

Updated home page to use the same navsubjects partial as the sidebar and deleted all evidence of my incompetency

Removed about partial, which was actually a duplicate of pages partial

Added two nice right-facing carets to the search bar so people get a sense they are actually being productive.

In a glorious show of decisiveness, spliced the get started link back out again

Fixed the incorrect path for the png fallback for the ol logo. Bloody equal signs are going to be the death of me one day.

Moved a lot of the sidebar logic around using partials

Changed input-xxlarge to input-block-level where possible, so we don't rely on an actual fixed width.

In the process of adding a dandy live search function to the staff database. Standby for magic.

Signed-off-by: Kenneth Lim <kenlimmj@gmail.com>